simplified project structure
