+++ tests/functional/pages/fab_ui_build_profile_sector.py
