+++ tests/functional/pages/fab_ui_case_study_images.py
