+++ tests/functional/pages/fab_ui_confirm_identity.py
