+++ tests/functional/pages/fab_ui_edit_sector.py
