+++ tests/functional/pages/fab_ui_upload_logo.py
