+++ tests/functional/pages/fas_ui_creative_industry.py
