+++ tests/functional/pages/fas_ui_health_industry_summary.py
