+++ tests/functional/pages/fas_ui_industries.py
