+++ tests/functional/pages/fas_ui_tech_industry.py
