+++ tests/functional/pages/sso_ui_logout.py
