+++ tests/functional/pages/sso_ui_password_reset.py
