+++ tests/functional/pages/sso_ui_register.py
