+++ tests/functional/steps/fab_when_def.py
@@ -48,9 +48,9 @@ from tests.functional.features.steps.fab_when_impl import (
     select_random_company,
     sso_change_password_with_password_reset_link,
     sso_go_to_create_trade_profile,
-    sso_sign_in,
     sso_open_password_reset_link,
     sso_request_password_reset,
+    sso_sign_in,
     sso_supplier_confirms_email_address
 )
